+++ compiler/src/main/java/info/teksol/mc/messages/ERR.java
@@ -151,6 +151,7 @@ public class ERR {
     public static final String LVALUE_VARIABLE_EXPECTED = "Variable expected.";
     public static final String METHOD_CALL_UNDEFINED = "Unknown method '%s'.";
     public static final String METHOD_CALL_UNRESOLVED = "Cannot resolve method '%s'.";
+    public static final String MISSING_MLOG_KEYWORD_PREFIX = "An mlog keyword is expected (if '%s' is intended as an mlog keyword, use ':%<s' instead).";
     public static final String MISSING_MODULE_DECLARATION = "Module declaration not found in source file.";
     public static final String MLOG_BLOCK_CONFLICTING_LABEL = "Label '%s' conflicts with a declared input or output variable.";
     public static final String MLOG_BLOCK_DUPLICATE_LABEL = "Label '%s' is already declared in this mlog block.";