+++ compiler/src/main/java/info/teksol/mc/messages/WARN.java
@@ -6,15 +6,12 @@ import org.jspecify.annotations.NullMarked;
 public class WARN {
     public static final String ABSOLUTE_ADDRESSING = "Unable to meet the requirement for symbolic links. Some portions of the program still depend on absolute addressing.";
     public static final String DEPRECATED_USE_OF_REQUIRE_REMOTE = "Using the 'remote' directive after 'require' is deprecated. Move the 'remote' directive in front of 'require'.";
-    public static final String DEPRECATED_USE_OF_REMOTE = "Using `remote` to mark variables/functions for export is deprecated. Use `export` instead.";
     public static final String FORMAT_PRECLUDED_BY_STRING_LITERAL = "A string literal precludes using 'format' instruction for print merging.";
     public static final String FUNCTION_NO_LONGER_SUPPORTED = "Function '%s' is no longer supported in Mindustry Logic version %s; using '%s' instead.";
     public static final String LINKED_UNKNOWN_TYPE_SPEC = "Unknown linked block type specification '%s' (must use block type, e.g. '@memory-cell').";
     public static final String LINKED_VARIABLE_NOT_RECOGNIZED = "Linked variable name '%s' doesn't correspond to any known linked block name.";
     public static final String LITERAL_LOSS_OF_PRECISION = "Loss of precision while creating mlog literal (original value %s, encoded value %s).";
     public static final String LITERAL_UNSAFE_DECIMAL_RANGE = "Literal '%s' exceeds safe range for integer operations (0 ... 2**53).";
-    public static final String MISSING_MLOG_KEYWORD_PREFIX = "Using mlog keywords without the ':' prefix is deprecated (use ':%s' instead).";
-    public static final String MISSING_MODIFIER_PARENS = "Modifier '%s' requires parentheses around its arguments.";
     public static final String MISSING_PRINTFLUSH_ADDED = "Automatically added missing 'printflush(message1);' call to the end of the program.";
     public static final String OPTIMIZATION_PASSES_LIMIT_REACHED = "Optimization passes limit (%d) reached.";
     public static final String PRINTF_NOT_ENOUGH_ARGUMENTS = "The 'printf' function doesn't have enough arguments for placeholders: %d placeholder(s), %d argument(s).";