+++ compiler/src/test/java/info/teksol/mc/mindcode/compiler/antlr/MindcodeParserTest.java
@@ -274,9 +274,8 @@ class MindcodeParserTest extends AbstractParserTest {
                     external var $x = 5, $y;
                     mlog("abc") var abc;
                     mlog("abc" + "def") var abc_def;
-                    remote a = 10;
-                    external p1 var b;
-                    external p1 mlog("c") var d;
+                    external(p1) var b;
+                    external(p1) mlog("c") var d;
                     linked cell1, message1;
                     
                     allocate stack in cell1;